refactor: move dataset types from atoms to types file
